Move tree view search to tree view code
